Move Footer into different directory and update file imports that use the Footer
